requirements.txt
cent
django-cors-headers